src/main/java/com/syncretis/rest_training/service/user/UserDetailsServiceImpl.java
package com.syncretis.rest_training.service.user;

import com.syncretis.rest_training.model.User;
import lombok.AllArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.stereotype.Service;

@Service
@AllArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {
    private final UserService userService;

    @Override
    public UserDetails loadUserByUsername(String username) {
        User user = userService.findUserByName(username);
        return UserDetailsImpl.getUser(user);
    }
}
